// src/types/MicrorregioesType.ts
export type Microrregioes =
  | "microrregioesParaiba"
  | "microrregioesMato Grosso do Sul"
  | "microrregioesMato Grosso"
  | "microrregioesMaranhao"
  | "microrregioesGoias"
  | "microrregioesEspirito Santo"
  | "microrregioesDistrito Federal"
  | "microrregioesCeara"
  | "microrregioesBahia"
  | "microrregioesAmazonas"
  | "microrregioesAmapa"
  | "microrregioesAlagoas"
  | "microrregioesAcre"
  | "microrregioesMinas Gerais"
  | "microrregioesPara"
  | "microrregioesParana"
  | "microrregioesPernambuco"
  | "microrregioesPiaui"
  | "microrregioesRio de Janeiro"
  | "microrregioesRio Grande do Norte"
  | "microrregioesRio Grande do Sul"
  | "microrregioesRondonia"
  | "microrregioesRoraima"
  | "microrregioesSanta Catarina"
  | "microrregioesSao Paulo"
  | "microrregioesSergipe"
  | "microrregioesTocantins";